Move dungeon party travel adapters into domain assembly
